Search is now its own path. Seperate page for search, not part of FindArticles anymore. Larger search bar when searching.
