+++ src/articles/FindArticles.js
@@ -6,21 +6,15 @@ import Interests from "./Interests";
 import SearchField from "./SearchField";
 import * as s from "./FindArticles.sc";
 import LoadingAnimation from "../components/LoadingAnimation";
-import Search from "../components/Search";
+import useQuery from "../hooks/useQuery";
 
 import ExtensionMessage from "./ExtensionMessage";
 import LocalStorage from "../assorted/LocalStorage";
 
 import ShowLinkRecommendationsIfNoArticles from "./ShowLinkRecommendationsIfNoArticles";
-import { useLocation } from "react-router-dom";
 import { APIContext } from "../contexts/APIContext";
 import useExtensionCommunication from "../hooks/useExtensionCommunication";
 import { getPixelsFromScrollBarToEnd } from "../utils/misc/getScrollLocation";
-// A custom hook that builds on useLocation to parse
-// the query string for you.
-function useQuery() {
-  return new URLSearchParams(useLocation().search);
-}
 
 export default function NewArticles() {
   const searchQuery = useQuery().get("search");
@@ -158,8 +152,6 @@ export default function NewArticles() {
     });
   }
 
-  console.log('Search:', searchQuery);
-
   return (
     <>
       <ExtensionMessage
@@ -170,12 +162,10 @@ export default function NewArticles() {
         setDisplayedExtensionPopup={setDisplayedExtensionPopup}
       ></ExtensionMessage>
 
-      {searchQuery === null && (
-        <Interests
-          api={api}
-          articlesListShouldChange={articlesListShouldChange}
-        />
-      )}
+      <Interests
+        api={api}
+        articlesListShouldChange={articlesListShouldChange}
+      />
       
       <s.SortAndSearch>
         <SearchField api={api} query={searchQuery} />
@@ -185,13 +175,6 @@ export default function NewArticles() {
           setArticleList={setArticleList}
         />
       </s.SortAndSearch>
-
-      {searchQuery && articleList.length > 0 &&(
-        <Search
-          api={api}
-          query={searchQuery}
-        />
-      )}
 
       {articleList.map((each, index) => (
         <ArticlePreview
@@ -208,10 +191,6 @@ export default function NewArticles() {
           onArticleClick={() => handleArticleClick(each.id, index)}
         />
       ))}
-
-      {searchQuery && articleList.length === 0 && (
-        <>No articles found that match your search</>
-      )}
 
       {!searchQuery && (
         <ShowLinkRecommendationsIfNoArticles
